backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
supabase>=2.0.0
gunicorn>=21.2.0
shapely
pyproj
google-generativeai>=0.8.0
httpx
beautifulsoup4
aiohttp
python-multipart
email-validator
web3
py-solc-x
geopy
langgraph
langchain-google-genai
tensorflow
pillow
numpy
opencv-python-headless
requests
requests-oauthlib
reportlab
imageio
pymongo
cloudinary
twilio
pandas
scikit-learn
xgboost
bcrypt
